feat: [feature/community] - 필터링 수정, AccessLevel 처리, 게시글 렌더링 버그 수정, 프로필 이미지 적용, 레이아웃 개선
